+++ ugo/stop_waiting.go
@@ -1,20 +1,20 @@
 package ugo
 
 type stopWaitingManager struct {
-	largestLeastUnackedSent uint32
+	largestLeastUnackedSent uint64
 	state                   bool
 }
 
-func (s *stopWaitingManager) GetStopWaitingFrame(force bool) uint32 {
+func (s *stopWaitingManager) GetStopWaitingFrame(force bool) uint64 {
 	if s.state == true {
 		s.state = false
 		return s.largestLeastUnackedSent
 	}
-	
+
 	return 0
 }
 
-func (s *stopWaitingManager) SetBoundary(p uint32) {
+func (s *stopWaitingManager) SetBoundary(p uint64) {
 	s.largestLeastUnackedSent = p + 1
 	s.state = true
 }
