+++ ugo/utils/packet_interval.go
@@ -3,6 +3,6 @@ package utils
 // PacketInterval is an interval from one PacketNumber to the other
 // +gen linkedlist
 type PacketInterval struct {
-	Start uint32
-	End   uint32
+	Start uint64
+	End   uint64
 }
